Adapt p1 f1 schemas for energy float values

+++ cchloader/models/f1.py
@@ -8,12 +8,12 @@ class F1Schema(Schema):
     measure_type = fields.Integer(position=1) # hauria de ser sempre 11
     datetime = fields.DateTime(position=2, format='%Y/%m/%d %H:%M:%S')
     season = fields.Integer(position=3, validate=OneOf([0, 1]))
-    ai = fields.Integer(position=4)
-    ao = fields.Integer(position=5, allow_none=True)
-    r1 = fields.Integer(position=6, allow_none=True)
-    r2 = fields.Integer(position=7, allow_none=True)
-    r3 = fields.Integer(position=8, allow_none=True)
-    r4 = fields.Integer(position=9, allow_none=True)
+    ai = fields.Float(position=4)
+    ao = fields.Float(position=5, allow_none=True)
+    r1 = fields.Float(position=6, allow_none=True)
+    r2 = fields.Float(position=7, allow_none=True)
+    r3 = fields.Float(position=8, allow_none=True)
+    r4 = fields.Float(position=9, allow_none=True)
     reserve1 = fields.Integer(position=10, allow_none=True)
     reserve2 = fields.Integer(position=11, allow_none=True)
     source = fields.Integer(position=12,
